requirements.txt
boto3==1.28.71
botocore==1.31.71
et-xmlfile==1.1.0
greenlet==3.0.1
jmespath==1.0.1
mysql-connector-python==8.2.0
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
protobuf==4.21.12
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.23
tqdm==4.66.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
XML2Dict==0.2.2
xmltodict==0.13.0
